+++ diagnostics/diag3.csv
