+++ diagnostics/ship diagnostics.csv
