--- sun_boundary_v1.json
@@ -1,1 +1,0 @@
-[[0, 31, 32, 33, 34, 34, 34, 35, 35, 35, 36, 36, 36, 37, 37, 37, 37, 38, 38, 38, 39, 39, 39, 39, 40, 40, 40, 40, 41, 41, 41, 42, 43, 43, 44, 44, 45, 45, 46, 47, 47, 48, 48, 48, 48, 49, 49, 49, 49, 49, 50, 50, 50, 50, 50, 51, 51, 51, 51, 52, 52, 52, 52, 52, 52, 53, 53, 53, 53, 53, 54, 54, 54, 54, 54, 54, 55, 55, 55, 55, 55, 55, 55, 55, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56], [0, 35, 36, 37, 38, 39, 114, 39, 40, 114, 39, 41, 42, 42, 43, 113, 114, 43, 44, 112, 43, 45, 47, 112, 44, 48, 49, 112, 45, 110, 111, 109, 109, 110, 50, 109, 51, 109, 108, 53, 107, 54, 55, 105, 106, 101, 102, 103, 104, 105, 98, 99, 100, 101, 102, 59, 97, 98, 99, 59, 61, 94, 95, 96, 97, 63, 65, 92, 93, 94, 65, 67, 69, 90, 91, 92, 63, 66, 71, 73, 76, 88, 89, 90, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87]]
