--- tk test.py
@@ -1,5 +1,0 @@
-from tkinter import *
-root = Tk() 							# Create the root (base) window 
-w = Label(root, text="Hello, world!") 	# Create a label with words
-w.pack() 								# Put the label into the window
-root.mainloop()
