+++ visuals/textures/triangles_indices.py
